app/requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configargparse==0.15.1
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==3.1.1
decorator==4.4.2
enum34==1.1.10
Flask==1.1.2
funcsigs==1.0.2
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.17
idna==2.10
importlib-metadata==1.7.0
ipaddress==1.0.23
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==5.0.0
msgpack==1.0.0
msgpack-python==0.5.6
numpy==1.16.6
packaging==20.4
pathlib2==2.3.5
pickleshare==0.7.5
Pillow==6.2.2
pluggy==0.13.1
prompt-toolkit==1.0.18
protobuf==3.12.4
py==1.9.0
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.5.2
pyparsing==2.4.7
pytest==4.6.11
pyzmq==19.0.2
requests==2.24.0
scandir==1.10.0
simplegeneric==0.8.1
six==1.15.0
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
win-unicode-console==0.5
zipp==1.2.0
zope.event==4.5.0
zope.interface==5.1.2
